requirements.txt
pyyaml
requests
curses
